+++ terraform/cloudtrail.tf
@@ -105,7 +105,7 @@ resource "aws_cloudtrail" "memex" {
   name                          = "${var.project_name}-trail"
   s3_bucket_name                = aws_s3_bucket.cloudtrail[0].id
   include_global_service_events = true
-  is_multi_region_trail         = false
+  is_multi_region_trail         = true
   enable_log_file_validation    = true
 
   tags = {